# api/docs/swagger.yaml
definitions:
...
  model.ErrValidation:
    properties:
      invalid_params:
        items:
          $ref: '#/definitions/model.InvalidParam'
        type: array
    type: object
  model.InvalidParam:
    properties:
      name:
        type: string
      reason:
        type: string
    type: object
...
  model.Order:
    properties:
      comments:
        type: string
      customer_number:
        type: integer
      order_date:
        $ref: '#/definitions/pgtype.Date'
      order_number:
        type: integer
      required_date:
        $ref: '#/definitions/pgtype.Date'
      shipped_date:
        $ref: '#/definitions/pgtype.Date'
      status:
        type: string
    type: object
  model.OrderDetail:
    properties:
      order_line_number:
        type: integer
      order_number:
        type: integer
      price_each:
        type: string
      product_code:
        type: string
      quantity_ordered:
        type: integer
    type: object
  model.Payment:
    properties:
      amount:
        type: number
      check_number:
        type: string
      customer_number:
        type: integer
      payment_date:
        $ref: '#/definitions/pgtype.Date'
    type: object
...
  pgtype.Date:
    properties:
      infinityModifier:
        type: integer
      status:
        type: integer
      time:
        type: string
    type: object
info:
  contact:
    email: Martsoftfilmz@gmail.com
    name: Martsoft Inc
  description: This is the API E-Commerce businesses.
  title: Martsoft Inc E-Commerce API 2.0
  version: "1.0"
paths:
...
  /orderDetails/delete:
    delete:
      description: Delete order details
      parameters:
      - description: order_number mandatory
        in: query
        name: order_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Delete order details
      tags:
      - Orders
  /orderDetails/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrderDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve all OrderDetails
      tags:
      - OrderDetails
  /orderDetails/get_orderDetail:
    get:
      description: Fetch order details
      parameters:
      - description: order_number mandatory
        in: query
        name: order_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrderDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve order details
      tags:
      - OrderDetails
  /orderDetails/update:
    patch:
      description: updates OrderDetail
      parameters:
      - description: order_number
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/model.OrderDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrderDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Updates a  OrderDetail
      tags:
      - OrderDetails
  /orders/orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve all Orders
      tags:
      - Orders
  /payments/delete:
    delete:
      description: deletes a Payment
      parameters:
      - description: customer_number mandatory
        in: query
        name: customer_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: deletes a  Payment
      tags:
      - Payments
  /payments/get:
    get:
      description: Fetch a Payment
      parameters:
      - description: customer_number mandatory
        in: query
        name: customer_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      tags:
      - Payments
  /payments/update:
    patch:
      description: updates a single Employee
      parameters:
      - description: payment_date
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/model.Payment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Updates a  Payment
      tags:
      - Payments
...
schemes:
- http
swagger: "2.0"